fix!: remove log reporter functions
The log reporters should not be changed during runtime.

Should only be configured during `config()`.

The same when during test, use `configForTest()`.
