+++ packages/log/ts/config.spec.ts
@@ -82,7 +82,7 @@ test('configure as development will emit warning', () => {
 })
 
 test('config.configured', () => {
-  expect(config.isLocked).toBeFalsy()
+    expect(config.isLocked).toBeFalsy()
   config({ mode: 'test' })
   expect(config.isLocked).toBe(true)
 })
