+++ packages/log/ts/index.ts
@@ -8,7 +8,6 @@ export * from './formatter/index.js'
 export * from './getLogger.js'
 export * from './logLevel.js'
 export * from './logLevelFn.js'
-export * from './logReporter.js'
 export * from './memory/index.js'
 export * from './shouldLog.js'
 export * from './suppressLogs.js'
